backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"

[project]
name = "epi-registry"
version = "0.1.0"
description = ""
authors = [
    { name = "YodZ0", email = "timur.g00gle@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-vvv"
asyncio_mode = "auto"
testpaths = "tests"
filterwarnings = [
    "ignore::DeprecationWarning"
]